--- mock_event.go
@@ -1,17 +1,0 @@
-package hsm
-
-type MockEvent struct {
-	id string
-}
-
-func NewMockEvent(id string) *MockEvent {
-	return &MockEvent{id}
-}
-
-func (e *MockEvent) ID() string {
-	return e.id
-}
-
-var mockStartEventId = "start"
-var mockSkipEventId = "skip"
-var mockEndEventId = "end"
