+++ transition_test.go
@@ -6,14 +6,9 @@ import (
 	"testing"
 )
 
-func NewMockService() (Service) {
-	logger, _ := zap.NewDevelopment()
-	return NewDefaultService(logger)
-}
-
 func TestNewExternalTransition(t *testing.T) {
 	mockState := NewMockState(NilState)
-	mockEvent := NewMockEvent(mockStartEventId)
+	mockEvent := NewMockEvent(mockStartEventID)
 
 	transition := NewExternalTransition(mockEvent, mockState, NopAction)
 	assert.NotNil(t, transition)
@@ -21,15 +16,15 @@ func TestNewExternalTransition(t *testing.T) {
 
 func TestExternalTransition_Execute(t *testing.T) {
 	mockState := NewMockState(NilState)
-	mockEvent := NewMockEvent(mockStartEventId)
-	mockService := NewMockService()
+	mockEvent := NewMockEvent(mockStartEventID)
+	logger, _ := zap.NewDevelopment()
 
 	transition := NewExternalTransition(mockEvent, mockState, NopAction)
-	assert.Equal(t, mockState, transition.Execute(mockService, mockState))
+	assert.Equal(t, mockState, transition.Execute(logger, mockState))
 }
 
 func TestNewInternalTransition(t *testing.T) {
-	mockEvent := NewMockEvent(mockStartEventId)
+	mockEvent := NewMockEvent(mockStartEventID)
 
 	transition := NewInternalTransition(mockEvent, NopAction)
 	assert.NotNil(t, transition)
@@ -37,15 +32,15 @@ func TestNewInternalTransition(t *testing.T) {
 
 func TestInternalTransition_Execute(t *testing.T) {
 	mockState := NewMockState(NilState)
-	mockEvent := NewMockEvent(mockStartEventId)
-	mockService := NewMockService()
+	mockEvent := NewMockEvent(mockStartEventID)
+	logger, _ := zap.NewDevelopment()
 
 	transition := NewInternalTransition(mockEvent, NopAction)
-	assert.Equal(t, mockState, transition.Execute(mockService, mockState))
+	assert.Equal(t, mockState, transition.Execute(logger, mockState))
 }
 
 func TestNewEndTransition(t *testing.T) {
-	mockEvent := NewMockEvent(mockStartEventId)
+	mockEvent := NewMockEvent(mockStartEventID)
 
 	transition := NewEndTransition(mockEvent, NopAction)
 	assert.NotNil(t, transition)
@@ -53,9 +48,9 @@ func TestNewEndTransition(t *testing.T) {
 
 func TestEndTransition_Execute(t *testing.T) {
 	mockState := NewMockState(NilState)
-	mockEvent := NewMockEvent(mockStartEventId)
-	mockService := NewMockService()
+	mockEvent := NewMockEvent(mockStartEventID)
+	logger, _ := zap.NewDevelopment()
 
 	transition := NewEndTransition(mockEvent, NopAction)
-	assert.Equal(t, NilState, transition.Execute(mockService, mockState))
+	assert.Equal(t, NilState, transition.Execute(logger, mockState))
 }
